+++ prboom2/src/e6y.c
@@ -980,9 +980,9 @@ void e6y_WriteStats(void)
     sprintf(str,
       "%%s - %%%dd:%%05.2f (%%%dd:%%02d)  K: %%%dd/%%-%dd%%%lds  I: %%%dd/%%-%dd%%%lds  S: %%%dd/%%-%dd %%%lds\r\n",
       max.stat[TT_TIME],      max.stat[TT_TOTALTIME],
-      max.stat[TT_ALLKILL],   max.stat[TT_TOTALKILL],   allkills_len,
-      max.stat[TT_ALLITEM],   max.stat[TT_TOTALITEM],   allitems_len,
-      max.stat[TT_ALLSECRET], max.stat[TT_TOTALSECRET], allsecrets_len);
+      max.stat[TT_ALLKILL],   max.stat[TT_TOTALKILL],   (long)allkills_len,
+      max.stat[TT_ALLITEM],   max.stat[TT_TOTALITEM],   (long)allitems_len,
+      max.stat[TT_ALLSECRET], max.stat[TT_TOTALSECRET], (long)allsecrets_len);
     
     fprintf(f, str, stats[level].map, 
       stats[level].stat[TT_TIME]/TICRATE/60,
